requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
mysql-connector-python==8.2.0
pymysql==1.1.0
redis==5.0.1
celery==5.3.4
pefile==2023.2.7
joblib==1.3.2
numpy==1.26.2
pillow==10.1.0
tensorflow==2.15.0
python-multipart==0.0.9
scikit-image==0.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
scikit-learn==1.3.2
pandas==2.1.3
requests==2.31.0

